chore: Updated configurations
