+++ docs/my-app/tests/test-helper.ts
@@ -1,5 +1,6 @@
 import { setApplication } from '@ember/test-helpers';
-import { start } from 'ember-qunit';
+import { setupEmberOnerrorValidation, start } from 'ember-qunit';
+import { loadTests } from 'ember-qunit/test-loader';
 import Application from 'my-app/app';
 import config from 'my-app/config/environment';
 import * as QUnit from 'qunit';
@@ -8,7 +9,6 @@ import { setup } from 'qunit-dom';
 setApplication(Application.create(config.APP));
 
 setup(QUnit.assert);
-
-start({
-  setupTestIsolationValidation: true,
-});
+setupEmberOnerrorValidation();
+loadTests();
+start();
