+++ docs/my-v1-addon/tests/test-helper.ts
@@ -1,14 +1,14 @@
 import { setApplication } from '@ember/test-helpers';
 import Application from 'dummy/app';
 import config from 'dummy/config/environment';
-import { start } from 'ember-qunit';
+import { setupEmberOnerrorValidation, start } from 'ember-qunit';
+import { loadTests } from 'ember-qunit/test-loader';
 import * as QUnit from 'qunit';
 import { setup } from 'qunit-dom';
 
 setApplication(Application.create(config.APP));
 
 setup(QUnit.assert);
-
-start({
-  setupTestIsolationValidation: true,
-});
+setupEmberOnerrorValidation();
+loadTests();
+start();
