+++ docs/my-v1-addon/tsconfig.json
@@ -14,7 +14,10 @@
       "my-v1-addon/test-support/*": ["addon-test-support/*"],
       "*": ["types/*"]
     },
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "types": [
+      "ember-source/types"
+    ]
   },
   "include": [
     "addon/**/*",
