+++ packages/embroider-css-modules/package.json
@@ -45,9 +45,7 @@
     "dist"
   ],
   "scripts": {
-    "build": "concurrently 'pnpm:build:*' --names 'build:'",
-    "build:js": "rollup --config",
-    "build:types": "glint --declaration",
+    "build": "rollup --config",
     "lint": "concurrently 'pnpm:lint:*(!fix)' --names 'lint:'",
     "lint:fix": "concurrently 'pnpm:lint:*:fix' --names 'fix:'",
     "lint:hbs": "ember-template-lint .",
@@ -56,9 +54,7 @@
     "lint:js:fix": "eslint . --fix",
     "lint:types": "glint",
     "prepack": "concurrently 'pnpm:build:*'",
-    "start": "concurrently 'pnpm:start:*' --names 'start:'",
-    "start:js": "rollup --config --watch --no-watch.clearScreen",
-    "start:types": "glint --declaration --watch",
+    "start": "rollup --config --watch",
     "test": "echo 'A v2 addon does not have tests, run tests in tests/embroider-css-modules'"
   },
   "dependencies": {
