+++ tests/my-v2-addon/config/ember-cli-update.json
@@ -3,7 +3,7 @@
   "packages": [
     {
       "name": "ember-cli",
-      "version": "6.0.0",
+      "version": "6.1.0",
       "blueprints": [
         {
           "name": "app",
