+++ tests/my-v2-addon/tests/test-helper.ts
@@ -1,5 +1,6 @@
 import { setApplication } from '@ember/test-helpers';
-import { start } from 'ember-qunit';
+import { setupEmberOnerrorValidation, start } from 'ember-qunit';
+import { loadTests } from 'ember-qunit/test-loader';
 import * as QUnit from 'qunit';
 import { setup } from 'qunit-dom';
 import Application from 'test-app-for-my-v2-addon/app';
@@ -8,7 +9,6 @@ import config from 'test-app-for-my-v2-addon/config/environment';
 setApplication(Application.create(config.APP));
 
 setup(QUnit.assert);
-
-start({
-  setupTestIsolationValidation: true,
-});
+setupEmberOnerrorValidation();
+loadTests();
+start();
